Introduce attribute type descriptors
Ideally, the attribute type descriptors should be registered alongside their corresponding attributes in the OnTopic Editor. By having them in the OnTopic Library, we created an awkward dependency between the Editor and core library which makes maintenance difficult.

Given this, we're moving these simple derivatives from the OnTopic Library to the OnTopic Editor. This is made possible by the OnTopic Library's `TopicFactory` now using the `DynamicTopicLookupService`, such that `Topic` derivatives don't need to be manually registered. As such, so long as a project maintains a dependency on the OnTopic Editor, these types will all be registered.