Remove legacy configuration support for `HtmlAttribute`
By removing the `GetConfigurationValue()` calls, we do away with the legacy ASP.NET Web Forms User Control style configuration support, which was previously maintained for backward compatibility. As part of this, we have also moved the default values from the view component to the attribute topic view model, which is a more logical place to store the defaults. This will also allow us, eventually, to migrate the attribute topic view model to an immutable, read-only object—and ideally a C# 9.0 `record`.

Note: Some of these defaults needed to be applied to parent types of `HtmlAttributeTopicViewModel`, such as `TextAreaAttributeTopicViewModel` and `AttributeDescriptorTopicViewModel`. These defaults are shared by most other view components, and remain logical defaults for these common view models.

Some of these properties are now deprecated and will be removed before the release of OnTopic Editor 5.0. That's acceptable. Until they are actually removed, they should still be treated as part of the interface since they remain functional.