Remove legacy configuration support for `TextAreaAttribute`
By removing the `GetConfigurationValue()` calls, we do away with the legacy ASP.NET Web Forms User Control style configuration support, which was previously maintained for backward compatibility. As part of this, we have also moved the default values from the view component to the attribute topic view model, which is a more logical place to store the defaults. This will also allow us, eventually, to migrate the attribute topic view model to an immutable, read-only object—and ideally a C# 9.0 `record`.

The default value for `Rows` of `5` conflicted with the default value for `Rows` of `20` that we assigned as part of the `HtmlAttribute` update (8746bec). I considered unifying these defaults, but they make sense for their respective targets. Given that, I've instead marked `Rows` as `virtual` in `TextAreaAttributeTopicViewModel` and overriden it on `HtmlAttributeTopicViewModels`.