Remove legacy configuration support for `TopicListAttribute`
By removing the `GetConfigurationValue()` calls, we do away with the legacy ASP.NET Web Forms User Control style configuration support, which was previously maintained for backward compatibility. As part of this, we have also moved the default values from the view component to the attribute topic view model, which is a more logical place to store the defaults. This will also allow us, eventually, to migrate the attribute topic view model to an immutable, read-only object—and ideally a C# 9.0 `record`.

As part of this, I also removed support for a legacy configuration setting, `AllowedKeys`, which is no longer supported as part of the modern strongly typed attribute configuration, nor exposed to the OnTopic Editor interface as part of the default configuration.