Removed legacy configuration helper methods
In the legacy ASP.NET Web Forms implementation of the OnTopic Editor, attributes were implemented as ASP.NET User Controls and configuration was stored as a set of key/value pairs that would be dynamically injected into the instantiation of that component. Support for this was maintained in OnTopic 4.0.0 in order to maintain backward compatibility with that style database.

With OnTopic 5.0.0, we'll be removing support for the legacy configuration format, under the assumption that all implementors are either using the new configuration format, or can easily migrate to it. The last eleven commits have removed the calls to these from the attribute view components; now that it's no longr used, we can remove it entirely.

Note: These were already marked as deprecrated, so should not be a surprise to any implementors that might still depend upon them.