Updated to alpha release of OnTopic 5.0.0 (w/ view models as records)
Updated to the latest alpha release of OnTopic 5.0.0, which migrates all topic view models to C# 9.0's new `record` type, which is read-only and immutable. This is a preferrable format for view models since they aren't intended to be modified after creation—while still allowing them to be easily modified as part of a cloning process in the rare case that they need to be updated (e.g., via the new `with` syntax).

This will break much of the current view models. Those will be updated as part of subsequent commits.