Updated derived view models to use C# 9.0 `record` type
As part of the latest preview of OnTopic 5.0.0 (233c6d7), the view models in `OnTopic.ViewModels` now use C# 9.0's new `record` type, which is an immutable and read-only. This is more memory efficient, and better articulates that view models are not usually expected to be modified after creation.

As part of this, all view models that derive from those classes also need to be converted to records. In the OnTopic Editor, that includes the `AttributeDescriptorTopicViewModel` and, thus, all of the attribute topic view model classes (e.g., `TextAttributeTopicViewModel`).

There were two cases where the view components needed to update the attribute topic view model after it had initially been mapped. This is a rare scenario, but it does pop up on occassion, particularly when there are defaults that rely on business logic that can't be assessed until the view component is initialized. In these two cases (`TopicReferenceViewComponent` and `HtmlViewComponent`), I updated to the new C# 9.0 `with` syntax, which allows a `record` to be cloned with a new value.

Finally, because the project is still using ASP.NET Core 3.x, instead of ASP.NET Core 5.x, it needs a placeholder `IsExternalInit` class in order to allow the `init` setters on the `record` types.