Migrate `AttributeViewModel` to use C# 9.0 `record` type
This required modifying all derived topics—i.e., all classes in the `OnTopic.Editor.AspNetCore.Models` and `OnTopic.Editor.Models.Components.ViewModels` namespaces—to also be records.

In addition, this also required refactoring a few view components to modify values using the initializer syntax, instead of relying on the property setter after the type had been initialized.

Notably, as some of these relied on the `AttributeViewModel.Value` as a fallback, that also needed to be independently evaluated prior to initializing the `AttributeViewModel` or derivative by pulling the value from the `EditingTopicViewModel.Attributes` collection.