Migrated `AttributeBindingModel` to use C# 9.0 `record` type
This change is a bit more debatable, since we will _always_ need to immediately create a copy of the initial `record` after it's initialized. That said, the initial record will be empty, and after that, for the bulk of its lifetime, this enforces that the value is immutable. I.e., by the time it gets to the actual `EditorController`, it should never need to be modified again.