Migrated `TopicQueryOptions` to use C# 9.0 `record` type
This is a bit more debatable, but as the `TopicQueryOptions` is almost exclusively initialized via reflection as part of ASP.NET Core's model binding, treating the `TopicQueryOptions` as a `record` doesn't cause any harm, and helps communicate that, as an options object, it shouldn't be modified after creation.

As part of this, I refactored the `TopicQueryService` to use an independent counter instead of treating `ResultLimit` as not only the initial value but also as a counter. While that worked, it mutated the original object in a way that could cause issues should that same object reference be reused in a subsequent operation.