Migrated `ContentTypeListViewModel` to use C# 9.0 `record` type
This isn't strictly necessary, but it helps reinforce that this should be a read-only object—though the (necessary) exception to that is the actual collection, which is the primary property of this view model. Nevertheless, this helps maintain consistency with the rest of our view models (now) being C# 9.0 `record`s.

Since this is the first record to be created with `init;` properties in the `OnTopic.Editor.AspNetCore` project, as opposed to the `OnTopic.Editor.AspNetCore.Models` project, it requires that the `IsExternalInit` stub be put into place. We can remove this when we migrate to .NET Core 5.0.