Fixed `ResultLimit` logic in the `TopicQueryService`
When migrating the `TopicQueryOptions` to a `record`, I refactored how the remaining results were counted by introducing a `remainingResults` counter instead of repurposing `TopicQueryOptions.ResultLimit` as a counter.

Unfortunately, when I did this, I missed that the call to the private `MapQueryResult()` method is in a loop, and, therefore, that value would get reset for each main branch. Further, the private `IsValidTopic` helper function continued to check the `ResultLimit` as though it was a counter. Whoops.

These two issues are fixed by introducing `remainingResults` in the main `Query()` entry point, and then passing it as a `ref` to the `MapQueryResult()`. That way, the same value remains shared across each "branch" of the mapping. I also pass that value to `IsValidTopic` so it's evaluating the correct counter.