Exposed warning if `IsFullyLoaded` isn't set to `true`
While a rare and unexpected scenario, this could cause confusion to editors. As such, to help set expectations, while also exposing potential data integrity issues, this error notifies editors that a `Topic` entity isn't in a fully loaded state, and helps communicate what the impact is.

As this is a rare scenario, this implementation doesn't differentiate between relationships or references, nor does the underlying data model make any attempt to determine what references weren't able to be loaded. That can be done by a developer in response to this warning.