Removed unused, legacy `DefaultAttributeTypeViewComponent`
This was originally used as a base class for other view components. This approach was error prone since view components don't support virtual `InvokeAsync()` methods. This has previously been marked deprecated, and thus necessitated that all view components use the same stock `InvokeAsync()` logic. The only view component that continued to use it was the `DisplayOptionsViewComponent`, which was deleted (6d5461e). As such, we can now get rid of the `DefaultAttributesTypeViewComponent` without any risk.