Deleted the `CssClass` property from `AttributeDescriptor`
This was previously used to style individual attributes. In practice, this was almost never used—and for good reason, since we generally want all attributes to be displayed in a consistent manner. But even if it were supported, this introduces issues with the proposed attribute plugin format we want to migrate towards, since it would require CSS classes be introduced in the common CSS, but referred to by the plugins.

(As is, we may still need to devise a plugin format for CSS, but we'll cross that bridge when we come to it.)

This has previously been marked deprecated.