Deleted the `ContentTypes` property for the `NestedTopicList` attribute
In OnTopic 4.0.0, this feature was replaced with the `PermittedContentTypes` relationship, which is an easier and more maintainable solution for marking what content types can be added to a nested topic list. We no longer need to maintain this legacy `ContentTypes` property, which was exclusively used as part of the legacy `DefaultConfiguration` (95458cb).

This has previously been marked deprecated.