Deleted the `RootTopicKey` property
The `RootTopicKey` provided a reference to the `Key` to be used for the root topic when performing searches. E.g., in the `QueryableTopicListAttribute`, it was used to determine where the search should begin, and in the `TopicReferenceAttribute` it was used to determine the parent from which the topic list should be displayed.

In both cases, this was replaced with the `RootTopic` property, which represents a strongly typed topic reference to a specific topic Id. This provides a fully-qualified reference which is stored in the database in a way that enforces data integrity. This is also loaded back into the topic graph as a full relationship, so it doesn't need to be determined by the editor.

These were both previously marked as deprecated.