Deleted the `Columns` property from `TextAreaAttribute`
The `Columns` property is an old HTML attribute that isn't usually too useful for managing layout (even though `Rows` continues to be). In general, we want the width to be managed by the editor itself, based on the overall layout and the current screen size. As such, it doesn't make sense to manage this on a per-attribute basis.

This has previously been marked deprecated.