Establish `OnTopic.Editor.AspNetCore.All` meta project
The purpose of the meta project is to provide a container for coordinating access not only the core `OnTopic.Editor.AspNetCore` project, but also ancillary plugin projects. It will be distributed as a meta package that implementors can depend on instead of individually registering e.g. each of the standard attributes.