Moved `StandardEditorComposer` into new meta project
The purpose of the new `All` meta project is to act as a coordinator between the main project and the plugins, so that the main project doesn't need to maintain a dependency on any individual plugins. The `StandardEditorComposer` is the heart of that registration process, so should live in the `All` project.