Moved the `EditorViewModelLookupService` to the `All` meta project
The purpose of the new `All` meta project is to act as a coordinator between the main project and the plugins, so that the main project doesn't need to maintain a dependency on any individual plugins. The `EditorViewModelLookupService` is key to that registration process, and so should live in the `All` project.