Moved view to shorter path location
The views for view components don't strictly need to be organized by `/Areas`; the view component will also search the general folders. So long as the views are namespaced by `/Editor`, they should be fairly specific to the OnTopic Editor, while this prevents the highly nested directory structure that's unwieldy to work with and navigate.