Establish new `Attributes` project for attribute types
Currently, attribute type definitions are distributed across multiple folders and projects. Some are in the main `OnTopic.Editor.AspNetCore` project, while other parts are in the `OnTopic.Editor.Models` project. Ideally, attribute types should be treated as plugins, with the main editor having no awareness of them at all. To facilitate that, we're establishing a new `OnTopic.Editor.AspNetCore.Attributes` project which will provide the main out-of-the-box templates used in the OnTopic Editor.