Moved the `StandardEditorComposer` to the `Attributes` project
The `StandardEditorComposer` is responsible for responding to requests for attribute view components. In order to do this, it needs to know where the view components live. As such, it should live in the `Attributes` project.

As part of this, I removed the condition in the `StandardEditorComposer` that artificially restricted it to types that are in the current assembly. Technically, that will be fine once everything is migrated, but it presents issues during the migration. Keeping it out will also prove useful if we choose to move some attributes out to their own individual projects, while still having them referenced from the `Attributes` project.