Move all views to new `Attributes` project
As part of this, each view was updated to reference its own namespace, since each attribute definition will be placed within a unique namespace (e.g., `OnTopic.Editor.AspnetCore.Attributes.TextAttribute`). This isn't an issue yet, but will be once the topic view models have been migrated over.

I also established a new `_ViewImports` and `_ViewStart` for the views folder. The `_ViewImports` is considerably shorter than the corresponding `_ViewImports` in the main `OnTopic.Editor.AspNetCore` project, as it doesn't need to worry about ancillary features of the editor, such as the import/export process.

Finally, these views are moved into the shorter `/Views/` folder instead of `/Areas/Editor/Views/`. The path is still deep, but it's at least a bit easier to work with. As this is still namespaced by the `Editor` controller directory, we expect this to be good enough to avoid name conflicts in most expected scenarios.