Move all attribute descriptors to new `Attributes` project
As part of this change, I renamed the attribute descriptors to use the `AttributeDescriptor` suffix. This is more intuitive since they are versions of attribut descriptors. It also avoids stepping on .NET's convention of only using the `Attribute` suffix for `[Attribute]` classes.

This will (temporarily) break the application, as the corresponding topic view models haven't yet been renamed. That will occur in a subsequent step.