Move all view components to new `Attributes` project
This is the last of the major moves required to consolidate all of the attribute type definitions into the new `OnTopic.Editor.AspNetCore.Attributes` project. As with the last few commits, the project remains broken, but will be fixed in the next couple of commits as we resolve the final references within, mostly, the lookup services. Almost there!