Replace `EditorViewModelLookupService` with dynamic lookup
In order to remain compatible with a distributed set of plugin assemblies, the `EditorViewModelLookupService` should instead be based on the `DynamicTypeLookupService`, with a condition to pull in classes whose name ends with `TopicViewModel`. We already have an out-of-the-box class that does that: the `DynamicTopicViewModelLookupService`. Given that, we can simply replace the `EditorViewModelLookupService` with a call to the `DynamicTopicViewModelLookupService`.