Separated `ContentTypeList` and `NestedTopicList` view models
Previously, the `ContentTypeListViewComponent` and the `NestedTopicListViewComponent` shared the same underlying `NestedTopicListAttributeDescriptorTopicViewModel`. Since teh `NestedTopicList` attribute has been migrated to the new `Attributes` project, and the main `OnTopic.Editor.AspNetCore` project deliberately does not maintain a dependency on that project, this is no longer possible. Given that, I've created a new `ContentTypeListAttributeDescriptorTopicViewModel`, and set the existing `NestedTopicListAttributeDescriptorTopicViewModel` to derive from it so that we're not duplicating otherwise identical code.