Updated topic mapping annotations
The `[Follow()]` attribute was renamed to `[Include()]` to better indicate that it's not recursive, and to align with Entity Framework's nomenclature. The `Relationships` enum was renamed to `AssociationTypes` to better express that it's not specific to relationships, and to establish a more intuitive nomenclature for the different types of associations that can occur in the topic graph. And, finally, `RelationshipType` was renamed to `CollectionType` to better express that these are mappings of collections, which may or may not actually be relationships.