Updated to release candidate of OnTopic Data Transfer 2.0.0
As with OnTopic 5.0.0, this is not the final release, but rather a release candidate intended to do final integration with. Once the final OnTopic Data Transfer 2.0.0 release occurs, we'll need to do a final update. That said, we're not _expecting_ any breaking changes from now until then, though it's _conceivable_ that we'll incorporate additional options.