Renamed `TranslateTopicPointers` to `TranslateLegacyTopicReferences`
In OnTopic 5.0.0, with the formalization of topic references, we're no longer using the term "topic pointers". As is, this was always used inconsistently, with other places referring to them as "topic references". This is now standardized, and thus the name should be updated.

However, the topic references this option is referring to are for the legacy handling, not the new topic references feature. To avoid confusion, this should be clearly stated. That's especially true since this is intended exclusively for backward compatibility.