Configured serialization to ignore null values
This was probably always a good idea, but it's especially useful now that the OnTopic Data Transfer library's `TopicData` data interchange class has a deprecated `DerivedTopicKey` property which is needed to import old JSON files, but is no longer used for generating new JSON files.

Without the `IgnoreNullValues` option, we would end up writing an empty `DerivedTopicKey` entry for every `TopicData` record, which is not only unnecessary, but also confusing given that we no longer use the `DeriedTopic` nomenclature. (The new `BaseTopic` property is now tracked via `TopicData.References`, just as any other topic reference.)

Once we migrate to .NET 5.0, there are additional attributes which allow us to map this field on import, while hiding it from serialization. As we want to maintain backward compatibility with .NET Core 3.x, and don't otherwise need enough features in .NET 5.0 to justify multi-targeting each of the libraries, we're keeping this as a deprecated public property for now. We'll revisit this in a future version, once it makes sense to target .NET 5.0.