Introduced `ScriptResource` and `ScriptCollection`
The `ScriptResource` and `ScriptCollection` derive from `ClientResource` and `ClientResourceCollection<T>`, and introduce an `InHeader` and `IsDeferred` property for determining _how_ and _where_ scripts should be exposed. They provide a strongly typed collection of scripts that should be registered for, and displayed in either the header or the footer (depending on the `InHeader` bit).

The `GetResources()` overload allows the `inHeader` or, optionally, `isDeferred` parameters to be specified, thus filtering the results.