Expose new `StyleSheetCollection` and `ScriptCollection`
Expose the new `StyleSheetCollection` and `ScriptCollection` on the `AttributeDescriptorTopicViewModel` as, respectively, the `StyleSheets` and `Scripts` properties. This will allow view models to represent client-side resources that need to be injected into the layout outside of the boundaries of the view component. These can be hardcoded into derivatives of the `AttributeDescriptor` view model, or they can be added by the view component logic itself—potentially even based on configuration logic (e.g., to select different themes).