Moved attribute-specific packages to `Attributes` project
Client-side resources that are specific to attributes should be packaged as part of the `OnTopic.Editor.AspNetCore.Attributes` package. To establish that, resources downloaded via npm should be migrated over to a new `packages.json` in the attributes projects. In addition, common packages needed by the build process will also need to be migrated.