Moved attribute-specific build tasks to `Attributes` project
With the attribute-specific client-side resources now being registered with npm in the `OnTopic.Editor.AspNetCore.Attributes` project, the `gulpFile.js` now needs to be forked between the two main projects, with attribute-specific items maintained in the `Attributes` copy.

Unfortunately, this introduces a lot of duplication of code in the `gulpFile.js`. In a future task, we'll likely want to find a way to centralize this code so that it can be shared between these two projects. Until that point, however, this establishes a framework for two distinct build processes, one for the editor interface, and another for the attribute plugins.