Moved attribute-specific resourcesto `Attributes` project
Client-side resources that are specific to attributes should be packaged as part of the `OnTopic.Editor.AspNetCore.Attributes` package. To establish that, style sheets and scripts maintained as part of the repository need to be moved to the `Attributes` project. This complements a previous commit which migrated attribute-specific npm dependencies (49cadea).