Moved `NestedTopicList.js` back to main editor project
While the `NestedTopicList.js` is, indeed, primarily intended to support the nested topic list, the modal architecture is actually built into the editor itself, since it requires deep integration with the layout and the buttons. To better abstract this from the specific implementation of nested topics, and to better clarify the purpose, it is renamed to `Modal.js`, and the description has been updated.