Removed legacy options for `DateTimeAttributeDescriptor`
With the migration to the HTML5 `datetime-local`, the format for the `date`, `time`, and `datetime-local` formats is fixed to `yyyy-mm-dd`, `hh:mm:ss`, and `yyyy-mm-ddThh:mm:ss` formats, respectively. As such, there's no need to collect `DateFormat`, `TimeFormat`, or `DateTimeSeparator` anymore.

If views need different formats, then they should map them to a `DateTime` object in the view model, and use .NET's `ToString()` formatting options.