Hide toolbar until it's repositioned
Currently, the toolbar starts out inside the main body area, and then is repositioned via JavaScript. That JavaScript is deferred, since we've deferred most of the client-side JavaScript—including jQuery, which it relies upon—for performance reasons.

Ideally, this would be resolved by updating the HTML and CSS such that the fixed toolbar is placed in the header area upfront, without relying on JavaScript to reposition it. But that's a more invasive change that doesn't make sense to pursue until we revisit the overall admin layout, which has a number of quirks.

Until then, as a quick fix, I'm defaulting to hiding the toolbar, and then showing it once the repositioning is completed. This isn't especially elegant, but it avoids the irritating CLS on load, while keeping the initial load time fast due to the deferred JavaScript.