Configure libraries to multi-target .NET Core 3.x, .NET 5
In order to allow implementations to fully take advantage of .NET 5 performance benefits, while maintaining compatibility with .NET Core 3.x, setting the library to multi-target both .NET Core 3.1 and .NET 5.

As part of this, I'm updating the host project to use .NET 5 exclusively. This also allows us to remove setting C# 9.0 as the language for the host project, since that's the default to .NET 5. (This can't be done for the other projects, since it's not the default for .NET Core 3.1.)