Suppressed duplicate content warnings (`NU5118`)
When packing a Razor Class Library (RCL) which targets multiple frameworks (in this case, .NET Core 3.1 and .NET 5), there is a bug where static files are perceived to be duplicates (dotnet/aspnetcore#27740). To mitigate that, the associated `NU5118` warning is suppressed, to help reduce unnecessary clutter in our build messages. (We don't expect this to otherwise be an issue, so feel comfortable suppressing it at the project level.)

While I was at it, I removed some legacy suppressions in the main `AspNetCore` project which are no longer required.