Use `init` for implicitly mapped properties
Implicitly mapped properties—i.e., properties whose name and type are the same on both the source `Topic` and the target model—must be settable in order for the `TopicMappingService` to correctly map them. This introduces a `CA2227`, however, for collections, since collections shouldn't be settable. This is easily resolved by simply using `init`, which still allows reflection to set the property value. This allows us to remove the `CA2227`.