Updated to the prerelease of OnTopic 5.0.0
Updated to OnTopic 5.0.0 beta 2. This introduces some breaking changes that are specific to OnTopic Editor, and which will be fixed in the next couple of commits.

As part of this, I removed unnecessary package references in the `Host` file, as many of these packages are inherited from `OnTopic.Editor.AspNetCore`.