Moved property overrides into constructor
With OnTopic 5.0.0, we're preferring the use of `protected init` accessors instead of `virtual` read-only properties for the purpose of overriding the default values from base classes. This effectively presents the same functionality, as far as we're concerned, but helps limit exposure to potential unintended uses or bugs that can be introduced by overriding a property (such as issues with polymorphism if not setting the `base` class). This only affects `AttributeDescriptor` and its derivatives—and, therefore, primarily impacts the OnTopic Editor. Easy update.