Remove legacy `AsRelationship` configuration
To treat the `TokenizedTopicList` as a relationship, simply set the `TokenLimit` to greater than 1; it will automatically be treated as a relationship. This was left over from OnTopic 4.0.0 where, by default, the list of `Topic.Id`s would be stored as an attribute. With OnTopic 5.0.0, we're ensuring that all topic associations are now stored as either `Topic.References` or `Topic.Relationships`. The only variable is whether it's a 1:1 or 1:n relationship.