Remove multi-targeting of .NET 5
In a previous release, we updated the OnTopic Editor to multi-target .NET 5 (50bb74d). That was premature. The primary driver of this was to ensure that the OnTopic Editor would take advantage of performance improvements in the .NET 5 runtime, assuming that the application was targeting .NET 5. It turns out this is unnecessary; the .NET 5 runtime will automatically be used, even if the Razor Class Library targets .NET Core 3.x, if the current application context targets .NET 5.

If, in the future, we want any .NET 5 specific capabilities then we can reintroduce multi-targeting so that we can provide conditional features. Until then, there's no need or benefit for the libraries themselves to multi-target.

Note: The `Host` project continues to target .NET 5, since that's the runtime we expect most implementations to be targeting moving forward.