Ensure nullable variables, parameters are marked as nullable
Quite a few variables and parameters allowed null—and even defaulted to null, especially in terms of optional parameters—but were not yet annotated as nullable. This resolved the CS8600 warning.

Note: Additional warnings will emerge later as we fix other code analysis warnings. We'll need to do a second pass of these.