Accounted for possible null reference assignments
Null reference assignments can occur when a nullable value is being assigned to a variable or property that should be annotated as nullable, similar to CS8600. It can also indicate that there should be a guard clause to ensure that the value isn't null at that point. These need to be handled with care. This resolves the CS8601 warning.