Accounted for passing possible null reference to non-nullable parameter
This is typically resolved by adding guard clauses. In some cases, we can reasonably assume that the value will not, in practice, ever be null, and can resolve using the null-forgiving operator (`!`). In a few cases, nulls are an acceptable input, and can be instead resolved by updating the target parameter or generic type reference to allow nulls. Resolved CS8604.