Reintroduced the `AttributeBindingModelBinder` as `public`
In a previous commit, the `AttributeBindingModelBinder` was marked as `internal`. It's unclear how this wasn't caught in initial testing, but this now results in an exception, due to it not having a `public` constructor. It turns out that the _class_ can still be marked as `internal`, but the _constructor_ must be `public` to satisfy the expectations.