Ensured `Key` defaults to `null` for new topics
Due to a bug in the revised attribute population logic, the parent topic's `Key` was getting set for `isNew` topics, thus preventing them from getting automatically assigned based on the `Title`. This exposed that the conditions for handling this were a bit convoluted, with nearly every condition having an `!isNew` clause, except for the last two. To fix that, I moved the two conditions targeting `isNew` to the top, and removed the `!isNew` checks in the remaining expressions. This not only resolves the bug, but also cleans up the logic a bit.