Provided handling for filtering by `ContentType`
The `IncomingRelationshipViewComponent` can, as with several other components, be optionally filtered by `AttributeKey`/`AttributeValue`. Often, that `AttributeKey` is `ContentType`. That worked fine when `ContentType` was, in fact, stored in `Topic.Attributes`. Now that it's been elevated to a "core attribute", that isn't possible. This was already addressed as part of the `TopicQueryService` used by most other scenarios where filtering is an issue, but was missed for the `IncomingRelationship` attribute type. This update fixes that issue by putting in a clause that checks, specifically, for an `AttributeKey` of `ContentType`, and sources the `attributeValue` for it separately.