Introduced `SourceLink` for GitHub to allow source debugging
We already generate `*.Symbols.Nupkg` on deployment, and publish them to Azure DevOps symbols repository. But we don't currently provide any source code for that. Since all of our source code is already public, and tightly versioned via git commits, this is pretty simple to accomplish using `SourceLink`. Now, when implementors debug their application, they can step into the OnTopic Editor as part of that debugging process, if appropriate. This will also make it easier for Ignia to identify and isolate bugs that may occur in customer implementations.