Restricted `StandardEditorComposer` to local libraries
Previously, the `StandardEditorComposer`'s `IsAssignableFrom()` would return true if the _any_ view component was passed in. That's not right. The `StandardEditorComposer` _only_ knows how to handle view components that are part of either `OnTopic.Editor.AspNetCore` or `OnTopic.Editor.AspNetCore.Attributes`. This used to be limited to `AspNetCore`, but that was inadvertantly removed as part of a previous commit (b01dd18c).