Introduced icon for the NuGet packages
By placing the icon in the root, explicitly setting it to be packed, and then referencing it as the `<PackageIcon />` in the `Directory.Build.props`, we can ensure that every project ends up with the same icon. For now, we're using the Ignia logo as the icon, and have set it to 128x128 per NuGet recommendations.

Note that in order to pack the file, the `<None />` reference needed to crawl back one directory, since the `<ItemGroup />` will be evaluated in context of each individual `csproj`. The `<PackageIcon />` doesn't require this, however, since the `<None />` reference will ensure it's in the root of the `*.nupkg` file.