Exclude npm `package.json` from NuGet package
By updating the `package.json` and `package-lock.json` files to `Pack=false`, we ensure they aren't copied to the NuGet package unnecessarily. These are _exclusively_ needed for the build process itself.

The SDK-style project will automatically add these as `<Content />` items _after_ the `Directory.Build.props` have run. As such, if we include them there, then they'll just get overwritten with the default configuration. If we update them in the `Directory.Build.targets`, however, they'll be correctly updated, since this is evaluated at the _end_ of the SDK-style project file.