Move removal of client-side mapping files to `Directory.Build.targets`
Now that we've established a `Directory.Build.targets` (ba09497), we can move the `<Content Remove />` elements from the individual `csproj` files to the `Directory.Build.targets` file.

We don't appear to be able to do an `Update` with `Pack="false"` for a glob pattern, but that's not a big deal since we aren't likely going to need to open the `map` files in Visual Studio.