Reintroduced a `EditorViewModelLookupService`
OnTopic Editor 4.x included an `EditorViewModelLookupService` for identifying view models associated with the OnTopic Editor. This was removed earlier in the OnTopic Editor 5.0.0 development process as part of the plugin support, since the original derived from the `StaticTypeLookupService` and hard-coded the names of the view models. To support plugins, we wanted this to be dynamic, and so could just use the out-of-the-box `DynamicTopicViewModelLookupService`.

That worked fine. But a problem with it is it includes _all_ view models from _all_ assemblies. That may be fine for many—even most—implementations. But others may wish for tighter control over what view models are included, and especially if they're overriding out-of-the-box view models.

To better support this, we're reintroducing the `EditorViewModelLookupService` as a dynamic service that loads only view models that derive from `ContentTypeDescriptorViewModel` or `AttributeDescriptorViewModel`—i.e., view models that the OnTopic Editor specifically needs access to in order to support attribute type plugins.

It's worth noting that the original `EditorViewModelLookupService` also derived from the `TopicViewModelLookupService`, thus enforcing a dependency on the `OnTopic.ViewModels` project. Customers may want that, but they also may not want it. As such, it was previously removed in preference for combining multiple `ITypeLookupService` implementations via the new `CompositeTypeLookupService`. As a result, implementations previously relying on `EditorViewModelLookupService` to _also_ deliver view models from the `TopicViewModelLookupService` will need to be updated to implement that via the `CompositeTypeLookupService`.