Disable automatic generation of NuGet package
To avoid unnecessarily increasing build time and adding clutter to the build directory, we should only generate NuGet packages when needed, instead of having it enabled by default for every build. This is how the rest of our projects are setup; the `OnTopic.Editor.AspNetCore` project was a bit of an outlier.