Addressed `IsHidden` discrepancy in `TopicViewModel`
In OnTopic 5.0.0 RC1, the `IsHidden` property has been removed from `TopicViewModel` as it's almost never needed by implementations, yet was added to every view model. It is, however, needed for the OnTopic Editor in order to determine if the `AttributeDescriptorViewModel` should be rendered or not. This is easily handled by removing it from the creation of `TopicViewModel` (in e.g. `IncomingRelationshipViewComponent`) and adding it to `AttributeDescriptorViewModel`